+++ lib/src/analyzers/lint_analyzer/rules/rules_factory.dart
@@ -1,6 +1,7 @@
 import 'models/rule.dart';
 import 'rules_list/always_remove_listener/always_remove_listener_rule.dart';
 import 'rules_list/avoid_border_all/avoid_border_all_rule.dart';
+import 'rules_list/avoid_collection_methods_with_unrelated_types/avoid_collection_methods_with_unrelated_types_rule.dart';
 import 'rules_list/avoid_dynamic/avoid_dynamic_rule.dart';
 import 'rules_list/avoid_global_state/avoid_global_state_rule.dart';
 import 'rules_list/avoid_ignoring_return_values/avoid_ignoring_return_values_rule.dart';
@@ -48,6 +49,8 @@ import 'rules_list/provide_correct_intl_args/provide_correct_intl_args_rule.dart
 final _implementedRules = <String, Rule Function(Map<String, Object>)>{
   AlwaysRemoveListenerRule.ruleId: (config) => AlwaysRemoveListenerRule(config),
   AvoidBorderAllRule.ruleId: (config) => AvoidBorderAllRule(config),
+  AvoidCollectionMethodsWithUnrelatedTypesRule.ruleId: (config) =>
+      AvoidCollectionMethodsWithUnrelatedTypesRule(config),
   AvoidDynamicRule.ruleId: (config) => AvoidDynamicRule(config),
   AvoidGlobalStateRule.ruleId: (config) => AvoidGlobalStateRule(config),
   AvoidIgnoringReturnValuesRule.ruleId: (config) =>